+++ src/main/java/com/example/deliveryapp/domain/order/controller/OrderController.java
@@ -3,9 +3,9 @@ package com.example.deliveryapp.domain.order.controller;
 import com.example.deliveryapp.domain.common.annotation.Auth;
 import com.example.deliveryapp.domain.common.dto.AuthUser;
 import com.example.deliveryapp.domain.order.dto.request.OrderMenuRequest;
-import com.example.deliveryapp.domain.order.dto.request.OrderRequest;
 import com.example.deliveryapp.domain.order.dto.request.OrderStateUpdateRequest;
 import com.example.deliveryapp.domain.order.dto.response.OrderResponse;
+import com.example.deliveryapp.domain.order.service.CartService;
 import com.example.deliveryapp.domain.order.service.OrderService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -20,6 +20,7 @@ import java.util.List;
 public class OrderController {
 
     private final OrderService orderService;
+    private final CartService cartService;
 
     // 전체 주문
     @PostMapping("/orders")
@@ -32,7 +33,7 @@ public class OrderController {
     @PostMapping("/stores/{storeId}/orders")
     public ResponseEntity<Void> addCart(
            @PathVariable Long storeId, @Auth AuthUser authUser, @Valid @RequestBody OrderMenuRequest request) {
-        orderService.addCart(storeId, authUser.getId(), request);
+        cartService.addCart(storeId, authUser.getId(), request);
         return new ResponseEntity<>(HttpStatus.OK);
     }
